+++ test/AuctionHouse/MockCallback.sol
@@ -47,7 +47,7 @@ contract MockCallback is BaseCallback {
         address seller_,
         address baseToken_,
         address quoteToken_,
-        uint96 capacity_,
+        uint256 capacity_,
         bool prefund_,
         bytes calldata callbackData_
     ) internal virtual override {
@@ -70,7 +70,7 @@ contract MockCallback is BaseCallback {
 
     function _onCancel(
         uint96 lotId_,
-        uint96 refund_,
+        uint256 refund_,
         bool prefunded_,
         bytes calldata callbackData_
     ) internal virtual override {
@@ -83,7 +83,7 @@ contract MockCallback is BaseCallback {
 
     function _onCurate(
         uint96 lotId_,
-        uint96 curatorFee_,
+        uint256 curatorFee_,
         bool prefund_,
         bytes calldata callbackData_
     ) internal virtual override {
@@ -106,8 +106,8 @@ contract MockCallback is BaseCallback {
     function _onPurchase(
         uint96 lotId_,
         address buyer_,
-        uint96 amount_,
-        uint96 payout_,
+        uint256 amount_,
+        uint256 payout_,
         bool prefunded_,
         bytes calldata callbackData_
     ) internal virtual override {
@@ -138,7 +138,7 @@ contract MockCallback is BaseCallback {
         uint96 lotId_,
         uint64 bidId,
         address buyer_,
-        uint96 amount_,
+        uint256 amount_,
         bytes calldata callbackData_
     ) internal virtual override {
         if (onBidReverts) {
@@ -155,8 +155,8 @@ contract MockCallback is BaseCallback {
 
     function _onClaimProceeds(
         uint96 lotId_,
-        uint96 proceeds_,
-        uint96 refund_,
+        uint256 proceeds_,
+        uint256 refund_,
         bytes calldata callbackData_
     ) internal virtual override {
         if (onClaimProceedsReverts) {
