+++ test/AuctionHouse/cancelAuction.t.sol
@@ -8,8 +8,8 @@ import {Auctioneer} from "src/bases/Auctioneer.sol";
 import {AuctionHouseTest} from "test/AuctionHouse/AuctionHouseTest.sol";
 
 contract CancelAuctionTest is AuctionHouseTest {
-    uint96 internal constant _PURCHASE_AMOUNT = 2e18;
-    uint96 internal constant _PURCHASE_AMOUNT_OUT = 1e18;
+    uint256 internal constant _PURCHASE_AMOUNT = 2e18;
+    uint256 internal constant _PURCHASE_AMOUNT_OUT = 1e18;
     uint32 internal constant _PAYOUT_MULTIPLIER = 50_000; // 50%
 
     bytes internal _purchaseAuctionData = abi.encode("");
